platform_sdl: Use platform-appropriate asset path
This will use the platform-appropriate path to load assets, when
available, or the path the binary lives in.

In other words, the CWD does not affect assets loading anymore.

This also intrinsically handles some platforms, like for example 3DS and
macOS...

... except that for macOS it would require building the `.app` correctly
and move the assets at the right location. For now it has been disabled.